requirements.txt
Flask==3.0.3
gunicorn==22.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
Werkzeug==3.0.3
itsdangerous==2.2.0
click==8.1.7
Jinja2==3.1.4
wtforms~=3.1.2
python-dotenv~=1.1.1
